Rename package and enable discovery client and zuul proxy
