+++ api-gateway/src/main/java/com/healthy/gym/api/gateway/ApiGatewayApplication.java
@@ -1,9 +1,13 @@
-package com.healthy.gym.apigateway;
+package com.healthy.gym.api.gateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 
 @SpringBootApplication
+@EnableDiscoveryClient
+@EnableZuulProxy
 public class ApiGatewayApplication {
 
     public static void main(String[] args) {
